+++ inst/examples/examples-plot_jags_densitydx.R
@@ -1,7 +1,7 @@
 
+
 data <- serodynamics::nepal_sees_jags_output
 
-# Specifying isotype and stratification for traceplot.
 plot_jags_dens(
                data = data,
                iso = "HlyE_IgA",