+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ## New features
 
+* Updating density diagnostic function (#71)
 * Changes to `run_mod()` (#79):
    - `jags.post` now optionally included in output, as specified by argument
    `with_post`